Move configure_cors to module file
