+++ htsget-search/src/storage/data_server.rs
@@ -10,14 +10,12 @@ use std::net::{AddrParseError, SocketAddr};
 use std::path::{Path, PathBuf};
 use std::pin::Pin;
 use std::sync::Arc;
-use std::time::Duration;
 
 use axum::http;
 use axum::Router;
 use axum_extra::routing::SpaRouter;
 use futures_util::future::poll_fn;
 use http::uri::Scheme;
-use http::{HeaderValue, Method};
 use hyper::server::accept::Accept;
 use hyper::server::conn::{AddrIncoming, Http};
 use rustls_pemfile::{certs, pkcs8_private_keys};
@@ -25,7 +23,6 @@ use tokio::net::TcpListener;
 use tokio_rustls::rustls::{Certificate, PrivateKey, ServerConfig};
 use tokio_rustls::TlsAcceptor;
 use tower::MakeService;
-use tower_http::cors::{AllowHeaders, AllowMethods, CorsLayer};
 use tower_http::trace::TraceLayer;
 use tracing::instrument;
 use tracing::{info, trace};
@@ -33,7 +30,7 @@ use tracing::{info, trace};
 use htsget_config::config::DataServerConfig;
 
 use crate::storage::StorageError::{DataServerError, IoError};
-use crate::storage::UrlFormatter;
+use crate::storage::{configure_cors, UrlFormatter};
 
 use super::{Result, StorageError};
 
@@ -188,7 +185,7 @@ impl DataServer {
   pub async fn serve<P: AsRef<Path>>(mut self, path: P) -> Result<()> {
     let mut app = Router::new()
       .merge(SpaRouter::new(&self.serve_assets_at, path))
-      .layer(Self::configure_cors(
+      .layer(configure_cors(
         self.cors_allow_credentials,
         self.cors_allow_origin,
       )?)
@@ -230,33 +227,6 @@ impl DataServer {
   /// Get the local address the server has bound to.
   pub fn local_addr(&self) -> SocketAddr {
     self.listener.local_addr()
-  }
-
-  /// Configure cors, settings allowed methods, max age, allowed origins, and if credentials
-  /// are supported.
-  fn configure_cors(cors_allow_credentials: bool, cors_allow_origin: String) -> Result<CorsLayer> {
-    Ok(
-      CorsLayer::new()
-        .allow_origin(
-          cors_allow_origin
-            .parse::<HeaderValue>()
-            .map_err(|err| DataServerError(format!("failed parsing allowed origin: `{}`", err)))?,
-        )
-        .allow_headers(AllowHeaders::mirror_request())
-        .max_age(Duration::from_secs(CORS_MAX_AGE))
-        .allow_credentials(cors_allow_credentials)
-        .allow_methods(AllowMethods::list(vec![
-          Method::GET,
-          Method::POST,
-          Method::PUT,
-          Method::DELETE,
-          Method::HEAD,
-          Method::OPTIONS,
-          Method::CONNECT,
-          Method::PATCH,
-          Method::TRACE,
-        ])),
-    )
   }
 
   fn rustls_server_config<P: AsRef<Path>>(key: P, cert: P) -> Result<Arc<ServerConfig>> {
@@ -313,7 +283,7 @@ mod tests {
   use std::str::FromStr;
 
   use http::header::{ACCESS_CONTROL_REQUEST_HEADERS, ACCESS_CONTROL_REQUEST_METHOD, ORIGIN};
-  use http::{HeaderMap, Method};
+  use http::{HeaderMap, HeaderValue, Method};
   use reqwest::{ClientBuilder, Response};
 
   use htsget_test_utils::util::generate_test_certificates;
